+++ processor/src/test/fixtures/bad_input/com/example/inject/InjectParameterizedTypeDaggerModel.java
@@ -1,0 +1,20 @@
+package com.example.inject;
+
+import arez.annotations.Action;
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+import java.util.concurrent.Callable;
+import javax.annotation.Nonnull;
+
+@ArezComponent( dagger = Feature.ENABLE )
+public abstract class InjectParameterizedTypeDaggerModel
+{
+  InjectParameterizedTypeDaggerModel( @Nonnull final Callable<String> action )
+  {
+  }
+
+  @Action
+  public void myActionStuff()
+  {
+  }
+}